add the ability to auto decompress the exefs
this pull request is based upon the issue i had where godmode9 did not actually decompress the code.bin file. if the code.bin is not decompressed you end up in a bootloop where sslc fails to load. This prevents that by requiring the exefs.bin file instead of the code.bin file by itself.